Add and Update mining chain sync functions and get functions for mining chain

+++ token/token_auth.go
@@ -130,9 +130,9 @@ func calcSHA256(targetHash string, maxNumber int) int {
 	return -1
 }
 
-func MaxTokenFromLevel(level int) int {
+func MaxTokenFromLevel(level int) uint64 {
 	val := TokenMap[level]
-	return val
+	return uint64(val)
 }
 func CreditsRequiredforLevel(level int) uint64 {
 	credits := CreditLevelMap[level]
@@ -152,7 +152,7 @@ func ValidateWholeToken(tokenDetials string) (int, int, bool, error) {
 		return -1, -1, false, err
 	}
 	tokenVal := -1
-	tokenVal = calcSHA256(tokenCountHash, MaxTokenFromLevel(tokenLevel))
+	tokenVal = calcSHA256(tokenCountHash, int(MaxTokenFromLevel(tokenLevel)))
 	if tokenVal == -1 {
 		return -1, -1, false, fmt.Errorf("token Count is invalid")
 	}
